+++ docs/changelog.md
@@ -2,7 +2,7 @@
 
 This changelog is used to track all major changes to WTFIX.
 
-## v0.10.0 (UNRELEASED)
+## v0.10.0 (2019-05-13)
 
 **Enhancements**